feat(observability): Wave 6 — conditions, events, JSON, OTel, cols
Wave 6 of the 2026-05-21 audit (items #13, #14, #15, #16, #17).

#17 PrinterColumn: @PrinterColumn annotations on KafkaCluster /
KafkaNodePool / KafkaPodSet / KafkaTopic / KafkaRbac / KafkaUI status
fields. `kubectl get kafkaclusters` now shows KAFKA + PHASE; `get
kafkatopics` shows phase + partitions + RF; etc.

#13 JSON logging + reconcile-id MDC: added quarkus-logging-json
extension; `quarkus.log.console.json=true` in the prod profile only
(dev/test stay plain for readability). New `infra/ReconcileContext`
AutoCloseable wraps every `reconcile()` body and sets `reconcile-id`,
`kind`, `namespace`, `name` in MDC — JSON logs promote each to a
top-level field, so log aggregation can group by reconcile-id. Wired
into all six reconcilers.

#16 K8s Events: new `infra/EventRecorder` emits Normal/Warning events
via the K8s Events API. Wired into KafkaClusterReconciler to emit on
phase transition (READY/FAILED). Best-effort — emit errors are
swallowed so they never break the reconcile. Visible via
`kubectl describe kafkacluster`.

#15 Conditions API: new `crd/Condition` POJO + `infra/ConditionUtil`
(preserves `lastTransitionTime` on no-change per K8s convention).
`KafkaClusterStatus.conditions[]` carries Available / Progressing /
Degraded alongside the bespoke `phase` field. Generic tooling can react
to conditions without knowing our enum.

#14 OpenTelemetry: added quarkus-opentelemetry extension.
`quarkus.otel.sdk.disabled=true` by default — operators turn it on by
overriding the property + setting an OTLP endpoint env var. Once
enabled, fabric8 HTTP / Kafka AdminClient / outbound REST calls are
automatically instrumented.

Incidental fixes:
- Dropped Wave 2's wildcard-CEL rule on KafkaRbacKafkaAccess.topics;
  K8s rejected it as exceeding the x-kubernetes-validations cost budget
  (the rule was never actually live). Wave 1's YAML emitter still
  prevents injection.
- KafkaTopicSpec.replicationFactor: primitive `short` -> `int`. The
  fabric8 CRD generator silently drops primitive `short` fields from
  the OpenAPI schema, so re-applying the CRD broke kafkatopic-test
  with "unknown field spec.replicationFactor".

275/275 unit tests pass; smoke e2e green.

+++ src/main/java/se/afshin/yavari/kafka/operator/apicurio/ApicurioKafkasqlSupport.java
@@ -194,7 +194,7 @@ public class ApicurioKafkasqlSupport {
                     ns, registry.getMetadata().getName(), partitions);
         }
         spec.setPartitions(partitions);
-        spec.setReplicationFactor((short) 3);
+        spec.setReplicationFactor(3);
         spec.setConfig(Map.of(
                 "cleanup.policy", "compact",
                 "min.insync.replicas", "2",
